Applied new logic for end game and new round

+++ src/models/leader_board.cairo
@@ -2,13 +2,13 @@
 struct LeaderBoard {
     #[key]
     id: u32,
-    len_players: u16
+    len_players: u32
 }
 
 #[generate_trait]
 impl LeaderBoardImpl of LeaderBoardTrait {
     #[inline(always)]
-    fn new(id: u32, len_players: u16) -> LeaderBoard {
+    fn new(id: u32, len_players: u32) -> LeaderBoard {
         LeaderBoard { id, len_players }
     }
 }
